Moved user info to a bigger scope
